Fix sqlcmd (#43)
* modified Metrics and deleted context for CPU and Memory

* renaming CPU_Context to MetricContext

* changed type for mem table

+++ src/DataTransfer/SQL_Commands/CREATE_BLANK_MEM_TABLE.sql
@@ -1,5 +1,5 @@
 DROP TABLE MEM_Data
 CREATE TABLE MEM_Data (
-"usage" FLOAT,
+"usage" BIGINT,
 "timestamp" DATETIME PRIMARY KEY
 )
